+++ crates/agent_servers/src/acp.rs
@@ -12,6 +12,7 @@ use serde::Deserialize;
 use settings::Settings as _;
 use task::ShellBuilder;
 use util::ResultExt as _;
+use util::process::Child;
 
 use std::path::PathBuf;
 use std::process::Stdio;
@@ -40,7 +41,7 @@ pub struct AcpConnection {
     default_mode: Option<acp::SessionModeId>,
     default_model: Option<acp::ModelId>,
     root_dir: PathBuf,
-    child: ProcessGroup,
+    child: Child,
     _io_task: Task<Result<(), acp::Error>>,
     _wait_task: Task<Result<()>>,
     _stderr_task: Task<Result<()>>,
@@ -95,7 +96,7 @@ impl AcpConnection {
         if !is_remote {
             child.current_dir(root_dir);
         }
-        let mut child = ProcessGroup::spawn(child, Stdio::piped(), Stdio::piped(), Stdio::piped())?;
+        let mut child = Child::spawn(child, Stdio::piped(), Stdio::piped(), Stdio::piped())?;
 
         let stdout = child.stdout.take().context("Failed to take stdout")?;
         let stdin = child.stdin.take().context("Failed to take stdin")?;
@@ -974,63 +975,5 @@ impl ClientDelegate {
             .get(session_id)
             .context("Failed to get session")
             .map(|session| session.thread.clone())
-    }
-}
-
-/// A wrapper around `smol::process::Child` that ensures all subprocesses
-/// are killed when the process is dropped by using process groups.
-struct ProcessGroup {
-    process: smol::process::Child,
-}
-
-impl std::ops::Deref for ProcessGroup {
-    type Target = smol::process::Child;
-
-    fn deref(&self) -> &Self::Target {
-        &self.process
-    }
-}
-
-impl std::ops::DerefMut for ProcessGroup {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.process
-    }
-}
-
-impl ProcessGroup {
-    fn spawn(
-        mut command: std::process::Command,
-        stdin: Stdio,
-        stdout: Stdio,
-        stderr: Stdio,
-    ) -> Result<Self> {
-        #[cfg(not(windows))]
-        {
-            util::set_pre_exec_to_start_new_session(&mut command);
-        }
-        #[cfg(windows)]
-        {
-            // TODO(windows): create a job object and add the child process handle to it,
-            // see https://learn.microsoft.com/en-us/windows/win32/procthread/job-objects
-        }
-        let mut command = smol::process::Command::from(command);
-        let process = command.stdin(stdin).stdout(stdout).stderr(stderr).spawn()?;
-        Ok(Self { process })
-    }
-
-    #[cfg(not(windows))]
-    fn kill(&mut self) -> Result<()> {
-        let pid = self.process.id();
-        unsafe {
-            libc::killpg(pid as i32, libc::SIGKILL);
-        }
-        Ok(())
-    }
-
-    #[cfg(windows)]
-    fn kill(&mut self) -> Result<()> {
-        // TODO(windows): terminate the job object in kill
-        self.process.kill()?;
-        Ok(())
     }
 }
